Fix: Nullpointer crash when scrolling detail activity
App was crashing when launching the detail view activity
and scrolling through the recipe steps and ingredients
in the ExpandableListView.  Also was using the incorrect
layout for recycled views as the app was scrolled through.

Instead of using two layout files, I consolidated to one
layout file and hid the views that I don't want for the
different layouts depending on the object data type.